Abstracted common points of the main menu to shared code, selected first game on app startup
