+++ Shared/ChessPartyApp.swift
@@ -14,9 +14,9 @@ struct ChessPartyApp: App {
     @SceneBuilder var body: some Scene {
         WindowGroup {
             #if os(iOS)
-            IOSMainMenu().environmentObject(settings)
+            IOSNavigationView().environmentObject(settings)
             #elseif os(macOS)
-            MacOSMainMenu().environmentObject(settings)
+            MacOSNavigationView().environmentObject(settings)
             #endif
         }
         #if os(macOS)
